Revamped Metadata, structure, and position watcher
Now each "map" should be an object, with everything on the map as its
children.
Metadata is added as a script section to each parent (you can right
click on the parents to add this)
Metadata script will be "disabled" by default.  This is intentional.
Each script is only enabled when changing maps, with the old one being
disabled again.
Fixed PlayerMovement to have fewer update calls, and fixed the walking
animation a bit.
On each step, MapPositionWatcher is updated, which determines what map
the player is on.
BGMs can now be stopped with a call, and when BGM_Player is destroyed it
stops its own audio.
MetadataSettings destroys and creates its children when disabled and
enabled, respectively.  This enforces changing of metadata effects, and
reduces the number of game objects in the scene.

Added the exterior of the Safari Zone as an example of multiple maps
connected.