Better Player Movement, and Terrain Tags
Player movement is much cleaner code, and works much better now.
Figured out how to do terrain tags.  Don't make the collider in unity,
make it in tiled, adding a box collider to the desired tile. Tiled will
put the collider on every instance of that tile.  The just attach the
TerrainTagChecker and specify the terrain type