Fixed Character Walk Cycles, and better scaling
Character Walk Cycle is perfect now, independent of frame rate.  Also
reimported and fixed the maps/charsets.
All maps should be imported with a 0.0096 scale, and charsets should be
set to 96 pixels per unity.  This helps reduce loss in quality when
scaling to the desired screen sizes.
Also fixed the splicing on the charsets.  No more wiggling when walking.